clothe-ai-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
google
google-genai
